feat: implement UpdateSale and improve logging in CreateSale and DeleteSale

+++ template/backend/src/Ambev.DeveloperEvaluation.Application/Sales/DeleteSale/DeleteSaleHandler.cs
@@ -1,15 +1,8 @@
-﻿using Ambev.DeveloperEvaluation.Application.Sales.CreateSale;
-using Ambev.DeveloperEvaluation.Application.Users.CreateUser;
-using Ambev.DeveloperEvaluation.Domain.Entities;
-using Ambev.DeveloperEvaluation.Domain.Repositories;
+﻿using Ambev.DeveloperEvaluation.Domain.Repositories;
 using AutoMapper;
+using FluentValidation;
 using MediatR;
 using Microsoft.Extensions.Logging;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Ambev.DeveloperEvaluation.Application.Sales.DeleteSale
 {
@@ -19,28 +12,26 @@ namespace Ambev.DeveloperEvaluation.Application.Sales.DeleteSale
     public class DeleteSaleHandler : IRequestHandler<DeleteSaleCommand, bool>
     {
         private readonly ISaleRepository _saleRepository;
-        private readonly IMapper _mapper;
-        private readonly ILogger<DeleteSaleHandler> _logger;
+        private readonly IValidator<DeleteSaleCommand> _validator;
 
-        public DeleteSaleHandler(ISaleRepository saleRepository, IMapper mapper, ILogger<DeleteSaleHandler> logger)
+        public DeleteSaleHandler(ISaleRepository saleRepository, IValidator<DeleteSaleCommand> validator)
         {
             _saleRepository = saleRepository;
-            _mapper = mapper;
-            _logger = logger;
+            _validator = validator;
         }
 
-        public async Task<bool> Handle(DeleteSaleCommand command, CancellationToken cancellationToken)
+        public async Task<bool> Handle(DeleteSaleCommand request, CancellationToken cancellationToken)
         {
-            var sale = await _saleRepository.GetByIdAsync(command.SaleId);
+            var validationResult = await _validator.ValidateAsync(request, cancellationToken);
+            if (!validationResult.IsValid)
+                throw new ValidationException(validationResult.Errors);
 
-            if(sale == null)
-            {
-                _logger.LogWarning("Sale {CustomerId} not found.", command.SaleId);
-                return false;
-            }
+            var sale = await _saleRepository.GetByIdAsync(request.SaleId, cancellationToken);
+            if (sale == null)
+                throw new KeyNotFoundException($"Sale with ID {request.SaleId} not found");
 
-            var deletedSale = await _saleRepository.DeleteAsync(sale, cancellationToken);
-            return deletedSale;
+            await _saleRepository.DeleteAsync(sale, cancellationToken);
+            return true;
         }
     }
 }
